requirements.txt
# Core dependencies (required)
PyPDF2>=3.0.0
numpy>=1.21.0

# Advanced ML dependencies (optional but recommended)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=2.0.0

# Additional utilities
pandas>=1.3.0

# For Gemini integration (optional)
google-generativeai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.9